// types/database.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      order_items: {
        Row: {
          id: string
          order_id: string
          price_snapshot: number
          product_id: string
          quantity: number
          seller_id: string
          title_snapshot: string
        }
        Insert: {
          id?: string
          order_id: string
          price_snapshot: number
          product_id: string
          quantity: number
          seller_id: string
          title_snapshot: string
        }
        Update: {
          id?: string
          order_id?: string
          price_snapshot?: number
          product_id?: string
          quantity?: number
          seller_id?: string
          title_snapshot?: string
        }
        Relationships: [
          {
            foreignKeyName: "order_items_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_items_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_items_seller_id_fkey"
            columns: ["seller_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      orders: {
        Row: {
          buyer_id: string
          created_at: string
          id: string
          status: string
          total: number
        }
        Insert: {
          buyer_id: string
          created_at?: string
          id?: string
          status?: string
          total: number
        }
        Update: {
          buyer_id?: string
          created_at?: string
          id?: string
          status?: string
          total?: number
        }
        Relationships: [
          {
            foreignKeyName: "orders_buyer_id_fkey"
            columns: ["buyer_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}